LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-focusing-steps-inert.html

<!DOCTYPE html>
<html>
<head>
<title>Test focusing steps when dialog is inert</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<input id="outer-input">
<dialog>
  <input autofocus>
</dialog>
<script>
function test_focusing_steps_with_inert_dialog(isModal) {
  const outerInput = document.querySelector("#outer-input");
  outerInput.focus();
  assert_equals(document.activeElement, outerInput,
    "Focus should be on element we just focused");

  const dialog = document.querySelector("dialog");
  assert_false(dialog.open, "Dialog should initially be closed");

  dialog.inert = true;

  if (isModal) {
    dialog.showModal();
    assert_equals(document.activeElement, document.body,
      "dialog.showModal(): focusing steps should apply focus fixup rule when dialog is inert");
  } else {
    dialog.show();
    assert_equals(document.activeElement, outerInput,
      "dialog.show(): focusing steps should not change focus when dialog is inert");
  }

  // Clean up
  dialog.close();
  dialog.inert = false;
}

test(() => {
  test_focusing_steps_with_inert_dialog(false);
}, "dialog.show(): focusing steps should not change focus when dialog is inert");

test(() => {
  test_focusing_steps_with_inert_dialog(true);
}, "dialog.showModal(): focusing steps should apply focus fixup rule when dialog is inert");
</script>
</body>
</html>
