LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-focusing-steps-inert-expected.txt



PASS dialog.show(): focusing steps should not change focus when dialog is inert
PASS dialog.showModal(): focusing steps should apply focus fixup rule when dialog is inert
